requirements.txt

litellm
python-dotenv
tweepy
snscrape
